권나윤/발주취소작업완료/20250304

+++ src/main/java/com/beauty1nside/purchs/service/purchaseService.java
@@ -27,5 +27,5 @@ public interface purchaseService {
 	public void purchUpdate(PurchUpdateVO vo);
 	
 	 // 발주 취소 메서드
-    boolean cancelPurchase(int companyNum, int purchaseNum);
+    boolean cancelPurchase(int companyNum, Long purchaseNum);
 }
